All files are structured properly
